backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis and Caching
redis==5.0.1
hiredis==2.2.3

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==43.0.3
pyotp==2.9.0
qrcode[pil]==7.4.2

# AI and ML
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2
weaviate-client==3.25.3
sentence-transformers==2.2.2
tiktoken==0.5.2

# Data Processing
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2

# HTTP and API
httpx==0.25.2
aiohttp==3.9.1

# Background Tasks
celery==5.3.4

# Monitoring and Logging
structlog==23.2.0
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Environment and Config
python-dotenv==1.0.0

# File Processing
python-magic==0.4.27
PyPDF2==3.0.1

# Date and Time
python-dateutil==2.8.2

# Email
aiosmtplib==3.0.1
jinja2==3.1.2

# Image Processing
Pillow==10.1.0

# Geospatial
geopy==2.4.1